requirements.txt
# Image Recognition Prototype - Requirements

torch
torchvision
numpy
faiss-cpu
pillow
fastapi
uvicorn
